[bazel_sdk] Prioritize HEAD variant over arch binaries
Use prebuilts for the "HEAD" API level from "variants" if they are
present. This makes it possible to use the appropriate prebuilts when
they are built, which could include via developer override in-tree or
when building the full SDK once support for "HEAD" when it is included
in the IDK.

Previously, the "binaries" metadata (for prebuilts built at "PLATFORM")
was processed second, meaning it would replace any metadata that existed
for "HEAD" in "variants". If both were present, this would cause the SDK
to use the "PLATFORM" prebuilts when targeting "HEAD" instead of the
prebuilts built at "HEAD".

There is no change in behavior since the IDK does not currently
generate prebuilts for "HEAD" and related metadata in "variants".

Bug: 385406226
Change-Id: Ic158743556ba3726ee66201fd8dd58a047b23f09
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1200490
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: Chase Latta <chaselatta@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>