[fxfs] Reduce the size KeyManager futures
KeyManager::get_key and the fscrypt variants were taking a future that
would fetch the wrapped keys if the keys weren't already cached. The
fetching future is typically more than 2KB. get_key moves the fetching
future into an Option which requires space for a second copy of the
fetching future in get_key's future. KeyManager::get_keys takes the
fetching future out of the Option to await it which moves into
get_keys's future. get_key awaits get_keys so get_key's future has space
for get_keys's future making get_key's future 3x the size of the
fetching future.

Changing get_key and the fscrypt variants to take an AsyncFnOnce instead
of the future stops the tripling of the size of the future. The
AsyncFnOnce is typically only 16 bytes and the Option of it is also only
16 bytes. Space for the fetching future is only necessary inside of
get_keys where it's created.

This reduces the typical size of KeyManager::get_key's future from 7344
bytes to 2656 bytes.

Bug: b/393365596
Change-Id: I32941580a674681c0b5fce6ff4e94cc8ff5586a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198688
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>