[bindgen] Control --explicit-padding with a knob
With the linked bug, we are in a case where we are using bindgen to
generate bindings that are indirectly (via bssl-sys) used by third-party
library (openssl). The latter code expected the construction of structs
not to have padding fields to initialize, and so we have a need to
control the generation of explicit padding fields. We default to 'on'
though, as this is generally useful for self-documenting ABI.

Bug: 394116035
Change-Id: Ie1fbded9ccf2718070cdc32dbc8d898f738b4d1b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1201306
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>