[fuchsia.io] Un-mark DeprecatedClone as `removed=HEAD`
We still have some load-bearing functionality that depends on the
existence of this method, so we should not mark it as removed until that
is fixed. Additionally, this is preventing building the IDK at HEAD.

For now we just mark the method as deprecated instead of removed.

Bug: 394878428
Test: No functional change.
Change-Id: I7024be42e63b3b03fd8d102acababd2b69eb20be
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1201227
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Brandon Castellano <bcastell@google.com>
API-Review: Chris Suter <csuter@google.com>
Reviewed-by: David Dorwin <ddorwin@google.com>
Reviewed-by: Chris Suter <csuter@google.com>