[overnet][usb] Retry sync packet on garbage
If garbage data is received on the usb channel when attempting to
connect, attempt to re-send the sync packet to restart the connection
dialog. For some reason the first read packet is getting eaten in the
usb_rs layer, so without this it doesn't synchronize.

The missing first packet should be fixed as well, but re-asserting the
state makes sense in this situation anyways. This fixes reconnection on
daemon restart, but there are still issues when the device goes away and
comes back in the higher overnet layers.

Bug: 393442313
Change-Id: I85e2407b21073d9c9817085f8513fb438e380c97
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1200492
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>