[build][grpc] Update build files to match v1.65.0
BUILD.gn generated locally with v1.65.0 checked out following the
instructions in //build/secondary/third_party/grpc/README.md, and then
copied over to the merge commit CL at https://fxrev.dev/1201532.

The patch file was also updated to generate a few, new, needed
libraries.

This roll is a blocker to the boringssl roll and v1.65.0 was chosen as
that was the first version tag that removed the code that the new
boringssl version does not like.

Bug: 339028859
Change-Id: I93d5cf0ed72ce72d2321df80eae1d676720b7a58
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1201685
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Anivia Li <aniviali@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>