[bt][ffx] Remove pandora from SDK partner_internal
The unstable FIDL library "fuchsia.bluetooth.pandora" was in the
"partner_internal" category because `ffx_bluetooth_tool` was in the
"partner" category due to the previous default for the template. Since
the tool is not in an IDK, remove the category from the dependencies.

This addresses one instance of an exception to "partner_internal"
stability rules.

Bug: 364727245
Change-Id: I18a1e2bdfb92d3e80e847b817629fd0040fbd775
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1194375
Fuchsia-Auto-Submit: David Dorwin <ddorwin@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
Reviewed-by: Ali Saeed <saeedali@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>