[Starnix] Resolve socket protocol on socket creation
Socket stores `protocol` field. Previously it was always set to the
value passed to `socket()` syscall during creation. The syscall allows
`protocol` parameter to be set to 0, which case the actual protocol
used by the socket is either TCP or UDP depending on the socket type
parameter. Updated `Socket::new()` to resolve the packet type, which
ensures the correct value is stored in the struct. This will allow
to pass correct protocol value to eBPF programs executed in Starnix.

Bug: 391380601
Change-Id: I950ab74cd3abba2ca45a000ebe651b4a0ba7f282
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1201768
Commit-Queue: Sergey Ulanov <sergeyu@google.com>
Reviewed-by: Benjamin Lerman <qsr@google.com>