[bus-pci] Split PCI bus driver for user and kernel implementations
This splits the two definitions of the PCI bus driver (kernel and user
space PCI) into separately named drivers that are included in the
configurations which need them. The driver using user space PCI support
is called 'bus-pci' and is included on all x86_64 and RISC-V
configurations as well as the vim3 and pkvm boards. The driver using
kernel PCI support is called 'bus-kpci' and is used on qemu-arm64.

Previously, the bus was always called 'bus-pci' and its definition was
determined by the platform_enable_user_pci GN argument.

This change modifies the 'arm64' and 'qemu-arm64' boards, both of which
are used in qemu to emulate an arm64 device.

Bug: 42182308
Change-Id: I738116ff8fa81c9332e522b675ae2c51843be861
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1199464
Reviewed-by: Mike Comfoltey <comfoltey@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: James Robinson <jamesr@google.com>