[kernel][thread] Construct free_dpc just before it's used
Move the construction of the free_dpc in Thread::Current::Exit to the
point at which its first used.  By moving this down, we eliminate a
blocker for making Dpc's fields const.

No behavior change.  This CL is the first in a series of small
improvements to the DPC service in preparation for resolving
b/395669867.

Bug: 395669867
Change-Id: Ide13685a166c651976c1f82be62ead42058b8cca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1203291
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Nick Maniscalco <maniscalco@google.com>