[fxfs] Box Item in async code
Item<ObjectKey, ObjectValue> is 224 bytes. Every time an Item is passed
to an async fn or held across an await, the future needs to capture the
Item and the future grows by 224 bytes. The MergerIterator::advance_impl
future is 1584 bytes because it holds 6 Items. Boxing Item here reduces
the size of the MergerIterator::advance_impl future to 312 bytes.

Bug: b/393365596
Change-Id: I08cd5542fccea472c8bd752e828679ff7bf2b38e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198689
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>