[fxblob] Reduce memory overhead of merkle trees
Open blobs keep their merkle trees in memory. These merkle trees account
for roughly half of fxfs' memory usage. The inner nodes of the trees can
be dropped since only the leaves are necessary to verify the data
blocks. The Vec holding the leaves is constructed with Vec::push which
has a growth factor of 2 which means that it could be twice the size
that it needs to be. Dropping the inner nodes and shrinking the Vec
holding the leaves reduces the overall merkle tree memory usage by 25%.

Bug: b/393365596
Change-Id: I3c3f105363349472d4b07b06575ae41ba1f1eb52
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1200346
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>