[display][coordinator] Switch empty config tests to color layer tests.
This CL removes the usage of empty (zero-layer) display configurations
from the Coordinator integration tests. This is a step towards removing
support for empty display configurations, for the reasons described in
the attached bug.

This CL introduces the use of color layers in the Coordinator
integration tests. The model is a bit confusing, due to an unfinished
migration from primary/color layers to unified layers in the Coordinator
<-> display client FIDL API. The model will be cleaned up by future
work, which will be easier to do once we remove empty configuration
support from the Coordinator.

Bug: 394413629
Test: fx test //src/graphics/display/drivers/coordinator
Change-Id: Ifc2918c573e696799ba9bb33c20cff7701a19a1f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1203366
Commit-Queue: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Reviewed-by: Yilong Li <liyl@google.com>