[botanist] Start checking for SSH key after sending serial commands.
Sometimes we see the serial commands to provision and cat the SSH key
get stalled for some time. However since we start reading for the
response before sending the commands, this may cause us to miss the
response because we time out before we can send it. Instead we should
start the timeout after sending the commands. The side effect is that in
the case that we hang completely, this will cause the timeout to write
to the serial socket to be 2 min instead of 10 seconds, multiplied by 5
tries.

Bug: 395706575
Change-Id: I29ce3a2400ab3e521724e575a077e2d6d55b0fed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1203870
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Carver Forbes <carverforbes@google.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>