[memory] Don't overcommit inspect slots
Before this change, logger.cc would allocate a fix amount of memory
for logging bucket names in the current configuration, which is
wasteful and does not print nicely (`ffx inspect show
core/memory_monitor` for instance would display a lot of empty
slots). This was done because the number of slots was not known at
logger instantiation time, but we can do better by deferring the
instantiation of the property to a time when the number of slots is
known.

Note that the number of buckets is expected to stay constant during
the entire lifetime of the logger, which is why we can get away with a
single dynamic allocation.

Change-Id: I0dcea98e872c9550561354fa05aece3a0e9802b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1203045
Reviewed-by: Étienne J. Membrives <etiennej@google.com>
Commit-Queue: Guillaume Chelfi <chelfi@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>