[ffx][fho] Removing re-export of ffx_writer
This removes re-exporting ffx_writer in fho. This
is done as part of refactoring the fho crate in order
to improve tool build times.

The next step is to refactor ToolIO, which is used as trait for
FfxMain, and is implemented by the Writers in ffx_writer.

Bug: 392174596
Change-Id: Ic2533b3c8fa1aaf4682beb87e5b182964ac90936
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1201770
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>