[diagnostics][fidl] Adjust string selector max length constant value
This value was based on the old fuchsia.component.types one. But this
is no longer an appropriate lower bound. We need to increase it and we
can increase it to the one defined in fuchsia.component.

Fixed: 395862030
Change-Id: I44d9e44373a4912907a4663e078247cac7c606fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1204114
API-Review: Christopher Johnson <crjohns@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Brian Bosak <bbosak@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>