[sestarnix] Add a `name` argument to `fs_node_init_on_create()`
SELinux `type_transition` rules include a form that is conditional
on the name of the new file node (i.e. the final element of its
initial path). This mechanism is mostly used to label `anon_inode`
instances, but is also supported for new file nodes.

Add a `name` argument to the init-on-create and can-create new file
hooks.

Bug: 385075470
Change-Id: Ib813da80e38add7b16f0fdc21133afba5d8717be
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1177812
Commit-Queue: Wez <wez@google.com>
Reviewed-by: Ambre Williams <ambre@google.com>