[display][coordinator] Stop applying empty config on client teardown.
Client:TearDown() currently unconditionally calls ApplyConfig() at the
end of its cleanup. This applies an empty (zero-layer) display
configuration, because the cleanup empties the layer lists for the
pending configuration.

This logic is incorrect, and conflicts with the correct behavior
implemented in Client::SetOwnership().

The correct behavior when a client disconnects is to have the
Coordinator reapply the last configuration applied by the new display
owner. If there is no such configuration, or if the configuration no
longer passes the verification, the current configuration must remain
displayed.

Bug: 394413629
Test: fx test //src/graphics/display
Change-Id: Ia393f778f97bdd7616517702d7d2064087cd6407
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1204347
Commit-Queue: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Reviewed-by: Yilong Li <liyl@google.com>