[starnix] Implement PTRACE_O_EXITKILL.
This CL implements the PTRACE_O_EXITKILL option in Starnix. When this
option is set, ptracees receive a SIGKILL when the ptracer terminates.

Note that this implementation is wrong, because PTRACE_O_EXITKILL
triggers when the ptracer exits, but starnix uses the thread group
leader as the ptracer. Some tests had to be skipped because of that.

Change-Id: Iee6ba15e1ab01b53ab0841f4f87d9cca33d0f806
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1201230
Reviewed-by: Theodore Dubois <tbodt@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>