[starnix] Drop out-of-date comment about priority
The patch to which the removed comment refers is
lore.kernel.org/linux-man/20220510220821.1481801-1-tbodt@google.com; it
was sent almost three years ago and accepted after just a few days (so
no need to continue to refer to it in the codebase as a thing that
perhaps-might-happen-if-it-hasn't-already-happened). To carry the load
carried by the earlier comment we add new comments on each of
`sys_getpriority` and `sys_setpriority` - it might seem like in so
doing we're swinging an overly-heavy hammer, but... the asymmetry is
really surprising to those coming to this area of the code and not
familiar with the subtlety (at least, that was how I felt 🙂).

We also tweak the names of the fields in the code itself to more
clearly communicate when we are dealing with "user-space" priorities
versus "raw" priorities - it was weird how `priority` was assigned once
in each regime (as a parameter in -20..19 "user-space" and as a local
field in 1..40 "raw").

Change-Id: Ic56e5dff3a879a8a40c879e3752970e07dd9e0b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198288
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Nathaniel Manista <nathaniel@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>