[pci] Check if device is extended (ecam) or not (cam)
This is needed in order to compute the address of device configurations.
crosvm presents a pci devices with a  "pci-host-cam-generic" compatible
string as opposed to qemu's pcie device with an "pci-host-ecam-generic"
compatible string. This uses the compatible string to decide if the
device is pci or pcie.

Bug: 42182308
Change-Id: If2173ef5483f1daff1fad2ddfd81d98d7388f1de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1204564
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: James Robinson <jamesr@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>