[resultdb] Preserve directory structure in artifact names.
This also adds an OutputDir field to the TestDetails struct so that we
can use that to write the per-test logs from tefmocheck into and we can
omit that dir from the artifact name since it should be common among all
the artifacts for a test. It also changes the paths in the OutputFiles
to be relative to the OutputDir, so fxrev.dev/1203828 needs to go in
first to update the recipes to get the correct paths if OutputDir is
provided.

Bug: 394639114
Change-Id: Ic2ff0eedcd6ea643a818fac197c834d18ccfa7b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1200111
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>