[fxfs] Link count for unlinked files with no links should be 0
Previously, this would report 1, but change to have it report 0.

In Fxfs, reference count refers to all internal references to an object.
So, even for objects living in the graveyard, the reference count is 0.
However, we've been using this reference count as the link count of a
file. This is not quite consistent with what link count refers to. For
example, with files that have been unlinked at there are no longer any
links to it, the link count should be 0 (Fxfs will previously report
link count as 1 as the object lives in the graveyard).

Add test in starnix syscall test suite to ensure consistency with Linux.

Bug: 391458506
Change-Id: I25010b6af60c4144c450e6c84f77ced49db6f9aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1202865
Reviewed-by: James Robinson <jamesr@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Rui Qi Sim <rqsim@google.com>