[stack-migration] "force NS2" is an enum variant
Encode the decision to force use netstack2 (as the result of failed
healthchecks in the previous boot), as a dedicated enum variant.

This allows us more explicitly track whether we're considering the
"forced setting" or ignoring it. Start ignoring it when publishing
changes in the desired netstack version to the rollback module.
Otherwise the rollback state machine may incorrectly enter `Canceled`
state.

Fixed: 395913604
Change-Id: Iefce52ad6f4ec06a07ff3ad5ac27b602de943509
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1203304
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Peter DeLong <pdelong@google.com>