[assembly] Unify sampler config in the diagnostics subsystem
Working towards using config capabilities instead of config_data for
Sampler configuration.

As part of that, we can move the work that sampler does to expand some
templates given the component_id_index values to assembly. To accomplish
this, we'll need access to the ComponentIdIndex in the Storage config
from the Diagnostics subsystem configuration.

This change moves this configuration of Sampler from Storage to the
diagnostics subsystem preparing for that change.

Bug: 395159893
Change-Id: I41677cdfe2ab50574adcc0d814ff56cfc8277740
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1203964
Commit-Queue: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>