[device-watcher] Open directories with PERM_READABLE
Previously in io1, the ability to enumerate directory entries and open
child entries was unprivileged. In the current io2 rights model, we want
to constrain these to the ENUMERATE and TRAVERSE/CONNECT rights,
respectively.

To ensure we can start to enforce this, we must ensure PERM_READABLE,
which is a convenient alias including these rights, is used when opening
directories that we want to perform these operations on.

Bug: 324111518
Change-Id: I8716ece193fe074984176e5f2f6be36c7dee8adf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1204165
Commit-Queue: Brandon Castellano <bcastell@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>