[ffxutil] Initial support for use_ffx_strict
Add a "strict" ffx invocation mode. This does not yet pass "--strict"
to ffx, since most botanist commands that run ffx do not yet support
the flag. But we _can_ pass configs on the command line instead of
via "ffx config set", and we can pass "--machine json" to commands.
Unfortunately, some commands also do not yet support "--machine", so
there is a mechanism to remove "--machine" where it doesn't work.

We are still starting the daemon, providing an isolateDir, etc -- those
will be necessary until it is possible to run all the ffx commands with
"--strict".

This change also reverts fxr/1192683, since it uses a different approach
for managing configs in invocations.

Bug: 391391857

Change-Id: I001f3aeb4ac3597b67824de1a2e10d3944bdd78c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1201784
Reviewed-by: Ina Huh <ihuh@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>