[display][coordinator] Improve integration test comments.
Thanks to a recent change in the Controller testing API, one of the
tests is no longer racy. This CL adds documentation to the testing API,
and removes the raciness warning.

Another test was very likely to report false positives. This CL greatly
reduces the odds of a false positive by adding a hard-coded timeout to
work around the lack of a testing API for event states.

Last, this CL renames a few variables and rewords a few comments, to
make the test logic easier to follow.

Bug: 388885807
Test: fx test //src/graphics/display/drivers/coordinator
Change-Id: Id2d1fb7bd13a569e9bc82ec7eb5077b01b1f3164
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1205566
Reviewed-by: Yilong Li <liyl@google.com>
Commit-Queue: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>