[kernel][vm] Virtualize VmObject lock
Changes VmObject to not assume that the lock it should use is in the
shared state, but rather requires the implementation provide the lock.
In the case of VmObjectPaged it further defers the lock to the linked
VmCowPages. The lock returned is ultimately the same lock from the
shared hierarchy state, but this enables it to be changeable.

Bug: 338300943
Change-Id: I68208735ccd2b1ddbff9437e90efed11131dcf00
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1204906
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>