[boringssl] Access test data as files
Following upstream's lead in the past, we previously had a scheme in
place to embed the contents of a large number of test data files and
access their contents at runtime through a generated function that knows
how to access these blobs. This file is unwieldy at 27MiB and upstream
has done away with it (and the way to regenerate it) at HEAD. We don't
need to do that to access test data though: we can define them instead
as `resource()`s that can be accessed through normal file I/O and make
the definitions of those table-driven and based of the list of sources
generated in BUILD.generated_tests.gni.

Bug: 339028859
Change-Id: Ie99bacdd3e857048bca051c267347a772e62fc64
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1202464
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Joshua Liebow-Feeser <joshlf@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>