[ffx][fho] Removing ffx_writer dependency in fho
This removes the dependency on ffx_writer in fho, allowing the
refactoring of TryFromEnv implementations for the Writers to move
to ffx_writer.

This also moves ToolIO trait to /src/developer/lib/writer since it is
used by on-device commands as well as ffx.

Finally, this adds the assert to fho and ffx_writer that they are
host toolchain only.

Bug: 394964930
Bug: 392174596
Change-Id: Icc53b94a9ea9076f723a0f2cfad06ab2b140c48f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1204351
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>