[sampler-config] Remove Optional selectors
This is another instance in which implementation details of sampler had
made it to the Configuration definitions. Instead of keeping vectors of
optional selectors, we can just remove selectors we don't need to track
anymore without impacting the configuration data types.

I realized this in the follow-up, but took the changes to this CL to
keep the follow-up change smaller.

Bug: 395172409
Change-Id: Ifadd17195c2649403ffd4281fa9838658ba02716
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1205846
Reviewed-by: Clayton McCray <claytonmccray@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>