[unwinder] Add async unwinding API
...which allows clients to inject Memory implementations which provide
data asynchronously, such as debuggers running on separate host from the
target process. Currently this is only used in tests, follow up changes
will add support in clients.

Tests are added to ensure that CFI based unwinding continues to work
from the eh_frame section. Testing debug_frame section integration
can be added later but is currently a taller task than it seems since it
requires a more complex test driver than currently exists for the
unwinder to provide the ability to pause and fetch memory from disk for
the unwinder's Memory APIs.

A new dependency on the zxdb MessageLoop implementation is added for
tests only. This is useful because it already supports both Fuchsia and
Linux platforms which the tests are already built to support.

Bug: https://fxbug.dev/292107156
Change-Id: Ia15310adb4ed27b685f60b57d9adbb5b2d7a1519
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1195596
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Adam Barth <abarth@google.com>