[power] Rename tests for shutdown-shim
Shutdown-shim is unrelated to power manager anymore, and so some tests
names are adjusted.

About half of the tests are redundant now as they are equivalent to the
other half. They can be removed in the next CL as the gerrit couldn't
show a good diff.

Bug: b/366035937
Test: CQ
Change-Id: I02a95e7f4cf58fc47f6847f2f676470551dc68ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1188554
Reviewed-by: Onath Dillinger <claridge@google.com>
Commit-Queue: Hai Bi <bihai@google.com>