[fxfs] Allow connections with writable rights to verified files
Remove restriction on *opening* verified files as writable. This does
**not** allow writing or resizing these files however, nor does it allow
any kind of operation that mutates the file's data.

While adding support in fxfs for the new fuchsia.io Open3 method, we
missed a check when opening a file that disallows opening as writable.
In practice writes still fail since we mark the handle as read-only.

Furthermore in Starnix, we keep handles to files cached, so we already
were keeping writable handles around. We enforce writability on handles
differently in Starnix as well. Starnix has used Open3, so it
effectively bypassed this check for quite some time.

We have additional access checks in Starnix that prevent opening a
verified file with O_RDWR or O_WRONLY, and additional checks in fxfs
that disallow write or resize operations.

Test: fx test fuchsia-pkg://fuchsia.com/fxfs-tests#meta/fxfs-platform-test.cm

Change-Id: I1d7dae6ffbde889d784c8c9aa8dea18d0d699afb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1206305
Reviewed-by: Nikita Jindal <nikitajindal@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>