[usb][devfs] Publishing usb-peripheral service
Updated the usb-peripheral driver to publish the service
instead of using devfs.

This is the last driver to use the usb-peripheral class
so this completes the migration of usb-peripheral from
devfs to aggregated services.

As part of cleanup, removing usb-peripheral from
class_names.h and devfs-driver.cml

Bug: 395911287

Testing: usb-virtual-bus-test

Change-Id: Ibe5b9f2381743524c2fd48bec24825a83c65388e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1200182
Commit-Queue: Garratt Gallagher <garratt@google.com>
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>