[component_manager] perform outgoing directory liveness checks
If a component is failing to read messages from its outgoing directory
channel, either because it is stuck, not implemented to, or simply
falling behind, the channel can fill up, crashing component manager and
along with it the system.

There's no way to query how many unread messages are in this channel, so
component manager will instead perform liveness checks. After every 1000
requests dispatched to a component's outgoing directory, component
manager will call `fuchsia.io.Directory/Sync` and wait to see if it
receives a response from the component within 3 seconds. If the response
to `Sync` does not arrive in this amount of time, a warning message
with the component's moniker is logged.

This will hopefully help debug scenarios where component manager crashes
due to a filled outgoing directory channel.

Bug: 393176930
Fixed: 297391405
Change-Id: I545ce7c2eafc219a6b94bb09084bf65e86a8595e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1204104
Reviewed-by: Amanda Tait <atait@google.com>
Commit-Queue: Claire Gonyeo <cgonyeo@google.com>