[assembly] Postprocessing script inputs are in config
The board config now lists all script dependencies
inside the config itself, as a new inputs = {} field.

This field maps the path relative to the execution
environment to the path inside the container.

This is necessary so that the board configs can be
converted to an AssemblyContainer which walks the
fields to copy files into the output directory.

Support for a postprocessing script is removed from GN,
because it is unused.

Bug: 388852272
Change-Id: I64a1fb99b73ce9a4cb936d0267bce87bd52996c4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1204506
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>