[build] Clean up errors with bazel directories
Previously is a directory was specified, but missing we'd get a log
message that did not look like an error, and gave the wrong message.

"""
INFO:

Non-directories are not allowed in --directory-outputs Bazel path, got:
"""

Now we prefix with 'Error:', and add a specific error for missing
directories.

"""
Error: Directory provided to --directory-outputs is missing, got:
"""

Change-Id: I8d6afa441be4f66aa913e640aa6573784854abf4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1207045
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>