[test][hrtimer] Improve the fidelity of the `Stop` method
The hrtimer API promises that a StartAndWait will return Canceled
for a stopped timer. The fake hrtimer used to not do that.

The new code tries to ensure that the cancelation process happens
in a way that is specified by the hrtimer API. The process is
less efficient than what hardware does, but should be compatible.

Multiply: starnix-tests
Bug: 373731551
Change-Id: I878ff0b34d8b2feb6ff94a74eab5821e29d8eecb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1206664
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Reviewed-by: Kyle Gong <kgong@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>