[kernel][dpc] Make Dpc's func_ and arg_ members const
Make Dpc's func_ and arg_ data members const in order to make it
easier to reason about the class's thread-safety.

Note, for this to happen, we had to remove a check from
test_dpc_queue.  That check was actually testing something that wasn't
guaranteed to be true (that it's safe to destroy/overwrite a Dpc once
it has begun executing, rather than before it has finished executing).

This is a refactoring change with no intended change in behavior.

Bug: 395669867
Change-Id: I72d80c2a3d332c034dabe1ac0d602234ab41a576
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1203446
Commit-Queue: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>