[kernel][vm] Do not suspend threads in kernel page fault
If a thread is blocked on a page request from a page fault in kernel
mode, let the page fault complete before processing a suspend signal.
This can only happen in the usercopy paths (both with and without
capture faults). We cannot safely resume from a suspend in this state
without risking data loss, since we do not immediately return out to
user mode, so hold off on the suspend altogether.

Test: Added pager core tests that fail without this change.
Bug: 42084841
Run-All-Tests: true
Change-Id: I3b533ba7be77d2413e462937bb89780b9fbfeb51
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1202225
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>