[kernel][vm] Re-order reference creation
Change the order of code in VMO reference creation so that fallible
checks are performed before constructing the VmObjectPaged. This is
necessary as destructing a partially constructed VmObjectPaged is
unsafe.

Comments and checks in ~VmObjectPaged are updated to reflect this, as
the existing check no longer makes sense as cow_pages_ is const and
assigned at construction.

Change-Id: I01b2968e0da4a6a34b12c7735851e1443956e87c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1206864
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>