[netstack3] Avoid dropping NDP watcher responder
In the NDP watcher implementation, flush_batch promises to check if a)
there's a batch of watch entries queued to send, and b) if there's
currently a pending watch request. However, it does not correctly check
a) before it decides to take() the pending watch request, which can
cause us to drop the responder (which clients of the ext crate end up
observing as a peer-closed).

Test: netstack3-bin-instrumented_test
Bug: 341759423
Change-Id: I682954622f8f98f27d511e5603af4c447265d4d3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1207485
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Commit-Queue: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Antonio Linhart <antoniolinhart@google.com>