[sampler][assembly] Prepare for soft migration of fire config schema
At the moment the sampler config looks like:

```
sampler: {
    metrics_configs: [paths],
    fire_configs: [paths],
}
```

In the follow-up we'll be using:

```
sampler: {
    project_configs: [paths],
    fire: {
        component_configs: [paths],
        project_templates: [paths]
    }
}
```

`metrics_configs` isn't used at the moment in any product.

`fire_configs` is used in products at the moment, but only passing a
`component_config`.

This change introduces the new schema for FIRE configs and handles the
component_configs part of it that is the one that we want to
soft-migrate.

The follow-up will delete the old schema entirely.

Bug: 395159893
Change-Id: Ie36e7935eef84f8ea8215c322281a600402d2b26
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1207544
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>