[kernel][vm] Simplify cow pages deferred deletion
Instead of using the generic VmDeferredDeleter for VmCowPages, relies
on the fact that VmCowPages have an explicit dead transition that is
invoked by the VmObjectPaged, and are not actually dealing with
recursion in their destructor.

Performing the dead transition loop explicitly in the ~VmObjectPaged
simplifies the VmCowPages::DeadTransition, as it does not need to
perform lock dropping, and saves some memory in the VmCowPages as it
does not need to have the potential to be in the deferred deletion
list.

Change-Id: I357766b4ef7d18b8a27d58bc0ef99cc690498714
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1206885
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>