[starnix] setpriority permissions
This fixes a Starnix crash that can be observed by starting a process,
setting the started process' RLIMIT_NICE to zero, and then calling
setpriority on the started process - the clamp fails to find an integer
that is below-or-at the rlimit of zero while also being at-or-above the
minimum raw priority of one.

From the reading and interactive experimenting I've been doing recently
it seems to be the case that the rlimit is only appropriate to use in
making the allow/disallow decision; the rlimit isn't appropriate to use
in the arithmetic calculation of the new priority value being set.

Bug: 322894197
Change-Id: I613bc8aee32921df760600505dba1b62b8c9abdb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1197195
Commit-Queue: Nathaniel Manista <nathaniel@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>